Detect file type for dotfiles & change background
autocmd is used to hard code the file type to certain file names, since
vim expects them to have the standard names starting with a dot but the
files are tracked in the repository without the leading dot.

Add a <leader> mapping to quickly toggle between the current and
previous file.

Add a <leader> mapping to quickly toggle dark/light background.

Cosmetic change: all key codes are lower case.