[vim] use standard Unicode in statusline (airline)
The fonts patched with Powerline symbols are buggy and limited. Using
standard Unicode provides good looks and requires no fiddling with
fonts, which is less headache in terms of portability. If support for
the used Unicode symbols turns out to be too limited, they can be
changed to normal ASCII characters.