[bash] fix inconsistent PATH and MANPATH
These two major problems existed:

 * PATH got multiple instances of the same entries for nested shells.
 * MANPATH got set, but ignoring the results of the system's `manpath`
   command. This caused some man pages to be missing, because the
   presence of MANPATH variable prevents man from looking up man pages
   at the paths returned from `manpath`.

Both issues are now fixed :)

BONUS: we abstract the last absolute locations we had in terms of where
Homebrew is installed. For Linux, Cygwin, or Mac without Homebrew we
assume all our custom tools will be installed under /usr/local. It is
easy enough to change that in case we have a different directory in a
specific system.