fix(tmux): changing background was ignored by tmux
When moving the configuration files to $XDG_CONFIG_HOME I forgot to
change the line in .bashrc that sources the tmux configuration so bash
was still trying to source ~/.tmux.conf, which no longer exists.

This commit fixes that. Additionally, the color setting code has been
factored out of tmux.conf into tmux-colors.conf (which can be sourced
from bash profile and tmux.conf). This way the rest of the code in
tmux.conf doesn't have to be executed again when changing the background
color.