bash: fix LS_COLORS
The solarize script evalss the output of the dircolors(1) but the
LS_COLORS variable gets exported in its environmnent, not the calling
shell where we actually need it so ls(1) and other programs wiill
inherit it.

Therefore, the evaluation of the dircolors(1) output is moved to the
change_bg() function in the shell, so that the LS_COLORS variable is
available to any children of the shell.

The reason why I had missed this is that in most systems I have ls
aliased to exa, which does not care about LS_COLORS and has its own
coloring system. On cygwin, however, exa is not available and I noticed
that the colors were missind; and indeed, on systems with exa the colors
are also missing if I run ls as \ls.