bash: factor out profile and aliases
This is a preparation to start experimenting with zsh. The idea is to
share the aliases and environment setup between both shells to avoid
having a bunch of duplication. Even if I decide to stick with zsh it
would be nice to keep the bash configs around for systems where I might
not want to install zsh for some reason. We'll see how this goes...