vim: always use dark background
The reason for this is that in some servers the automatic detection (see
:help background) is guessing it wrong and using a light background,
consequently burning my eyeballs every time I open vim.

Hard coding the dark background seems fine for now, as I seem to never
really have a need for the light version. If the needs arises in the
future, then I'll have to come up with a better solution, but KISS for
now.