zsh: prevent `tput` errors when $TERM is not set
When running a command via SSH without allocating a pseudo terminal,
for example, $TERM is not set, causing `tput` to freak out about not
knowing the terminal type.

So, we suppress the customization of man pages when $TERM is not set, in
which case we wouldn't have needed it anyway, as we don't even have a
terminal to type `man` :).