vim: fix loss of cursorline when using Telescope
The issue was that 'cursorline' is a window option but I was storing the
original value in global variable. So when Telescope (which has
'cursorline' not set in its window) entered insert mode, the InsertEnter
autocmd caused the global variable to store 0 (nocursorline). After
this, every InsertLeave event would use the global value and eventually
disable the cursorline everywhere.