vim: better macros to move lines
The old macro got triggered unintentionally all the time, especially
when leaving Insert mode. The reason was that one way terminals can
emulate <Alt> modifier key is by sending <Esc> rapidly followed by the
key that is supposed to be modified. So, when pressing <Esc> to leave
Insert mode and immediately pressing J to go down, my current line was
moved – not the intended behavior and incredibly annoying.