vim: add feedback for lua executor
This is useful when executing lines that produce no output, so I have
some confidence that something actually happened when pressing the key
bindings.

I am printing the message before executing the command because otherwise
this feedback would shadow the output of the command being executed
(which I am presumably interested in and would serve itself as
feedback).